Michaels branch (#19)
* Chapter 13 mastering the grammar

* Remove 13 placeholder.

* chapter 8 annotations

* Add requirements.

Be sure to run usethis::use_package("packageName") (in your console, not in the book itself) whenever you use a new package. That ensures that the system trying to build the notes has the packages you reference. Thanks!

* Delete extra chapter 8.

* Add showtext.

I meant to add this package before but didn't save, oops. I fixed another slight bug while I was at it and made sure it would render locally, so this SHOULD be it!

* Add janitor to requirements.

I *think* this is the last one.

Co-authored-by: Jon Harmon <jonthegeek@gmail.com>